feat: implement `Session::emitted_diagnostics`
Adds a way to get the emitted diagnostics as a `std::error::Error`.
`Session` (or `DiagCtxt`) must be built with a human buffer emitter
for this to work, as we need a human emitter to display the diagnostics.

The implementation is not great because we need quite a few workarounds for
the lack of APIs in anstream.