fix: windows eol lexing (#340)
Windows, unlike other OS, uses CRLF (`\r\n`) as end-of-line delimiter.
As the lexer relies on the end-of-line when handling `//`/`///` single
line comments, this creates a cross-platform inconsistent behavior (the
final `\r` ends up included in the comment parsed, see #315 ).

This PR adds a test for this hedge case and a suggested fix (reusing
memchr ability to match 2 different bytes in a given haystack)