perf(sema): better parallel parser scheduling (#428)
The current algorithm is batched:
1. parse all unparsed sources in parallel, waiting for all to finish
2. resolve all imports
3. add new sources to the list and repeat

Rewrite to be fully recursive:
1. spawn a parse job for all unparsed sources
2. each job resolves imports after parsing, spawning new jobs
recursively for new sources
3. top level function exits once all spawned jobs are done thanks to
`rayon::scope`

This process has to lock while adding sources to the list in 2. but
results it better usage of resources as new files don't have to wait for
the current batch to finish before they begin parsing and so on.

This is +~0% to +20%... varies a lot depending on input but always equal
or an improvement from what I've measured.